add ReferStyleDTO

+++ pixel-force-open-client/src/main/java/hiforce/pixel/open/client/param/result/ResultDTO.java
@@ -100,7 +100,7 @@ public class ResultDTO extends BaseParamDTO {
 
     public static long randomSeed() {
         Random random = new Random();
-        long num = random.nextLong();
+        long num = random.nextInt();
         return num > 0 ? num : -num;
     }
